added Room database, created Room entities, added MVVM architecture, and added Dagger Hilt dependency

+++ app/src/main/java/com/android/archives/ui/activity/TaskDetailViewActivity.kt
@@ -26,7 +26,7 @@ class TaskDetailViewActivity : AppCompatActivity() {
 
         intent?.let { it ->
             it.getIntExtra("task_id", 0).let {id ->
-                task = app.taskList.find { it.taskId == id }!!
+                task = app.taskList.find { it.taskId.toInt() == id }!!
             }
         }
